fix(suite-native): guard earn review screens against disconnected device
