+++ suite-native/module-earn/package.json
@@ -48,6 +48,7 @@
         "@suite-native/config": "workspace:*",
         "@suite-native/confirm-on-trezor": "workspace:*",
         "@suite-native/device": "workspace:*",
+        "@suite-native/device-authorization": "workspace:*",
         "@suite-native/device-manager": "workspace:*",
         "@suite-native/device-mutex": "workspace:*",
         "@suite-native/feedback-form": "workspace:*",
