+++ suite-native/module-earn/src/screens/yield/YieldDepositApprovalTransactionDataReviewScreen.tsx
@@ -5,6 +5,7 @@ import { type RouteProp, useRoute } from '@react-navigation/native';
 import { selectSelectedDevice } from '@suite-common/device';
 import { type YieldStackParamList, YieldStackRoutes } from '@suite-native/navigation';
 
+import { EarnDeviceConnectionGuard } from '../../components/earn/EarnDeviceConnectionGuard';
 import { YieldDepositApprovalReviewContent } from '../../components/yield/YieldDepositApprovalReviewContent';
 import { useYieldFlowData } from '../../hooks/yield/useYieldFlowData';
 
@@ -34,13 +35,15 @@ export const YieldDepositApprovalTransactionDataReviewScreen = () => {
     }
 
     return (
-        <YieldDepositApprovalReviewContent
-            approvalLimitType={approvalLimitType}
-            device={device}
-            flowData={flowData}
-            flowKey={flowKey}
-            transactionType={transactionType}
-            vaultTokenName={vaultTokenName}
-        />
+        <EarnDeviceConnectionGuard>
+            <YieldDepositApprovalReviewContent
+                approvalLimitType={approvalLimitType}
+                device={device}
+                flowData={flowData}
+                flowKey={flowKey}
+                transactionType={transactionType}
+                vaultTokenName={vaultTokenName}
+            />
+        </EarnDeviceConnectionGuard>
     );
 };
