+++ suite-native/module-earn/src/screens/yield/YieldDepositReviewScreen.tsx
@@ -11,6 +11,7 @@ import {
     YieldStackRoutes,
 } from '@suite-native/navigation';
 
+import { EarnDeviceConnectionGuard } from '../../components/earn/EarnDeviceConnectionGuard';
 import { YieldDepositReviewContent } from '../../components/yield/YieldDepositReviewContent';
 import { useYieldFlowData } from '../../hooks/yield/useYieldFlowData';
 import { buildYieldReviewPreview } from '../../utils/yield/yieldReviewOutputUtils';
@@ -76,5 +77,9 @@ export const YieldDepositReviewScreen = () => {
         return null;
     }
 
-    return <YieldDepositReviewContent flowData={flowData} flowKey={flowKey} preview={preview} />;
+    return (
+        <EarnDeviceConnectionGuard>
+            <YieldDepositReviewContent flowData={flowData} flowKey={flowKey} preview={preview} />
+        </EarnDeviceConnectionGuard>
+    );
 };
