+++ suite-native/module-earn/src/screens/yield/YieldDepositWrapReviewScreen.tsx
@@ -15,6 +15,7 @@ import {
     YieldStackRoutes,
 } from '@suite-native/navigation';
 
+import { EarnDeviceConnectionGuard } from '../../components/earn/EarnDeviceConnectionGuard';
 import { WrappedNativeTokenReviewContent } from '../../components/earn/WrappedNativeTokenReviewContent';
 import { useWrappedNativeReviewPreview } from '../../hooks/earn/useWrappedNativeReviewPreview';
 import { useYieldFlowData } from '../../hooks/yield/useYieldFlowData';
@@ -87,15 +88,17 @@ export const YieldDepositWrapReviewScreen = () => {
     }
 
     return (
-        <WrappedNativeTokenReviewContent
-            account={account}
-            amount={review.amount}
-            flowContext="in-flow"
-            flowType="wrap"
-            onBroadcast={handleBroadcast}
-            preview={preview}
-            spentToken={spentToken}
-            unsignedTransaction={review.unsignedTransaction}
-        />
+        <EarnDeviceConnectionGuard>
+            <WrappedNativeTokenReviewContent
+                account={account}
+                amount={review.amount}
+                flowContext="in-flow"
+                flowType="wrap"
+                onBroadcast={handleBroadcast}
+                preview={preview}
+                spentToken={spentToken}
+                unsignedTransaction={review.unsignedTransaction}
+            />
+        </EarnDeviceConnectionGuard>
     );
 };
